requirements.txt
flask
torch
transformers
pillow
streamlit

